Removing pointless media utils
