--- src/shared/utils/media/is-image.ts
@@ -1,5 +1,0 @@
-const imageRegex = /(http)?s?:?(\/\/[^"']*\.(?:jpg|jpeg|gif|png|svg|webp))/;
-
-export default function isImage(url: string) {
-  return imageRegex.test(url);
-}
