--- src/shared/utils/media/is-magnet-link.ts
@@ -1,9 +1,0 @@
-const magnetLinkRegex = /^magnet:\?xt=urn:btih:[0-9a-fA-F]{40,}.*$/;
-
-export default function isMagnetLink(url: string) {
-  return magnetLinkRegex.test(url);
-}
-
-export function extractMagnetLinkDownloadName(url: string) {
-  return new URLSearchParams(url).get("dn");
-}
