--- src/shared/utils/media/is-video.ts
@@ -1,5 +1,0 @@
-const videoRegex = /(http)?s?:?(\/\/[^"']*\.(?:mp4|webm))/;
-
-export default function isVideo(url: string) {
-  return videoRegex.test(url);
-}
